// src/api/generated/model/criterionKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export type CriterionKind = typeof CriterionKind[keyof typeof CriterionKind];


export const CriterionKind = {
  country_in: 'country_in',
  industry_in: 'industry_in',
  employees_between: 'employees_between',
  revenue_at_least: 'revenue_at_least',
  tag_in: 'tag_in',
} as const;